agent/pyproject.toml
[project]
name = "event-organizer-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "anthropic>=0.40",
    "agentmail",
    "python-dotenv",
    "pydantic[email]>=2.0",
    "modal",
    "fastmcp>=2.0",
    "fastapi[standard]",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
